dos: Change the DoS heartbeat line format
Fix a bug introduced in 94b56eaa7597e4a091a5b51d2c9032ea046631e3 which
overwrite the connection message line.

Furthermore, improve how we generate that line by using a smartlist and change
the format so it is clearer of what is being rejected/detected and, if
applicable, which option is disabled thus yielding no stats.

Closes #40308

Signed-off-by: David Goulet <dgoulet@torproject.org>